+++ src/main/java/agents/plantOperator/ConventionalPlantOperator.java
@@ -313,4 +313,10 @@ public class ConventionalPlantOperator extends PowerPlantOperator implements Fue
 			throw new RuntimeException(ERR_PAYOUT_VANISH + this);
 		}
 	}
+
+	@Override
+	protected double getInstalledCapacityInMW() {
+		return portfolio.getInstalledCapacityInMW(now());
+	}
+
 }